+++ weed/worker/tasks/balance/balance.go
@@ -80,51 +80,3 @@ func (t *Task) EstimateTime(params types.TaskParams) time.Duration {
 	// Could adjust based on volume size or cluster state
 	return baseTime
 }
-
-// Factory creates balance task instances
-type Factory struct {
-	*tasks.BaseTaskFactory
-}
-
-// NewFactory creates a new balance task factory
-func NewFactory() *Factory {
-	return &Factory{
-		BaseTaskFactory: tasks.NewBaseTaskFactory(
-			types.TaskTypeBalance,
-			[]string{"balance", "storage", "optimization"},
-			"Balance data across volume servers for optimal performance",
-		),
-	}
-}
-
-// Create creates a new balance task instance
-func (f *Factory) Create(params types.TaskParams) (types.TaskInterface, error) {
-	// Validate parameters
-	if params.VolumeID == 0 {
-		return nil, fmt.Errorf("volume_id is required")
-	}
-	if params.Server == "" {
-		return nil, fmt.Errorf("server is required")
-	}
-
-	task := NewTask(params.Server, params.VolumeID, params.Collection)
-	task.SetEstimatedDuration(task.EstimateTime(params))
-
-	return task, nil
-}
-
-// Register registers the balance task with the given registry
-func Register(registry *tasks.TaskRegistry) {
-	factory := NewFactory()
-	registry.Register(types.TaskTypeBalance, factory)
-	glog.V(1).Infof("Registered balance task type")
-}
-
-// Auto-register this task when the package is imported
-func init() {
-	factory := NewFactory()
-	tasks.AutoRegister(types.TaskTypeBalance, factory)
-
-	// Also register with types registry
-	tasks.AutoRegisterTypes(RegisterSimple)
-}
