+++ weed/worker/tasks/erasure_coding/ec.go
@@ -77,51 +77,3 @@ func (t *Task) EstimateTime(params types.TaskParams) time.Duration {
 	// Could adjust based on volume size or other factors
 	return baseTime
 }
-
-// Factory creates erasure coding task instances
-type Factory struct {
-	*tasks.BaseTaskFactory
-}
-
-// NewFactory creates a new erasure coding task factory
-func NewFactory() *Factory {
-	return &Factory{
-		BaseTaskFactory: tasks.NewBaseTaskFactory(
-			types.TaskTypeErasureCoding,
-			[]string{"erasure_coding", "storage", "durability"},
-			"Convert volumes to erasure coded format for improved durability",
-		),
-	}
-}
-
-// Create creates a new erasure coding task instance
-func (f *Factory) Create(params types.TaskParams) (types.TaskInterface, error) {
-	// Validate parameters
-	if params.VolumeID == 0 {
-		return nil, fmt.Errorf("volume_id is required")
-	}
-	if params.Server == "" {
-		return nil, fmt.Errorf("server is required")
-	}
-
-	task := NewTask(params.Server, params.VolumeID)
-	task.SetEstimatedDuration(task.EstimateTime(params))
-
-	return task, nil
-}
-
-// Register registers the erasure coding task with the given registry
-func Register(registry *tasks.TaskRegistry) {
-	factory := NewFactory()
-	registry.Register(types.TaskTypeErasureCoding, factory)
-	glog.V(1).Infof("Registered erasure coding task type")
-}
-
-// Auto-register this task when the package is imported
-func init() {
-	factory := NewFactory()
-	tasks.AutoRegister(types.TaskTypeErasureCoding, factory)
-
-	// Also register with types registry
-	tasks.AutoRegisterTypes(RegisterSimple)
-}
