+++ weed/worker/tasks/vacuum/vacuum.go
@@ -77,51 +77,3 @@ func (t *Task) EstimateTime(params types.TaskParams) time.Duration {
 	// Could adjust based on volume size or usage patterns
 	return baseTime
 }
-
-// Factory creates vacuum task instances
-type Factory struct {
-	*tasks.BaseTaskFactory
-}
-
-// NewFactory creates a new vacuum task factory
-func NewFactory() *Factory {
-	return &Factory{
-		BaseTaskFactory: tasks.NewBaseTaskFactory(
-			types.TaskTypeVacuum,
-			[]string{"vacuum", "storage"},
-			"Vacuum operation to reclaim disk space by removing deleted files",
-		),
-	}
-}
-
-// Create creates a new vacuum task instance
-func (f *Factory) Create(params types.TaskParams) (types.TaskInterface, error) {
-	// Validate parameters
-	if params.VolumeID == 0 {
-		return nil, fmt.Errorf("volume_id is required")
-	}
-	if params.Server == "" {
-		return nil, fmt.Errorf("server is required")
-	}
-
-	task := NewTask(params.Server, params.VolumeID)
-	task.SetEstimatedDuration(task.EstimateTime(params))
-
-	return task, nil
-}
-
-// Register registers the vacuum task with the given registry
-func Register(registry *tasks.TaskRegistry) {
-	factory := NewFactory()
-	registry.Register(types.TaskTypeVacuum, factory)
-	glog.V(1).Infof("Registered vacuum task type")
-}
-
-// Auto-register this task when the package is imported
-func init() {
-	factory := NewFactory()
-	tasks.AutoRegister(types.TaskTypeVacuum, factory)
-
-	// Also register with types registry
-	tasks.AutoRegisterTypes(RegisterSimple)
-}
